introduce AbstractIdentifiableDescriptor
