+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/descriptor/impl/AbstractIdentifiableDescriptor.java
@@ -1,0 +1,118 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.impl;
+
+import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.AdministrationDescriptor;
+import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.DescriptionDescriptor;
+import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.EndpointDescriptor;
+import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.IdentificationDescriptor;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+
+/**
+ * Methods for Identifiable.
+ */
+public abstract class AbstractIdentifiableDescriptor {
+
+    private String idShort;
+    private List<EndpointDescriptor> endpoints;
+    private AdministrationDescriptor administration;
+    private List<DescriptionDescriptor> descriptions;
+    private IdentificationDescriptor identification;
+
+    public AbstractIdentifiableDescriptor() {
+        idShort = null;
+        endpoints = new ArrayList<>();
+        administration = null;
+        descriptions = new ArrayList<>();
+        identification = null;
+    }
+
+
+    public String getIdShort() {
+        return idShort;
+    }
+
+
+    public void setIdShort(String idShort) {
+        this.idShort = idShort;
+    }
+
+
+    public List<EndpointDescriptor> getEndpoints() {
+        return endpoints;
+    }
+
+
+    public void setEndpoints(List<EndpointDescriptor> endpoints) {
+        this.endpoints = endpoints;
+    }
+
+
+    public AdministrationDescriptor getAdministration() {
+        return administration;
+    }
+
+
+    public void setAdministration(AdministrationDescriptor administration) {
+        this.administration = administration;
+    }
+
+
+    public List<DescriptionDescriptor> getDescriptions() {
+        return descriptions;
+    }
+
+
+    public void setDescriptions(List<DescriptionDescriptor> descriptions) {
+        this.descriptions = descriptions;
+    }
+
+
+    public IdentificationDescriptor getIdentification() {
+        return identification;
+    }
+
+
+    public void setIdentification(IdentificationDescriptor identification) {
+        this.identification = identification;
+    }
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        AbstractIdentifiableDescriptor that = (AbstractIdentifiableDescriptor) o;
+        return Objects.equals(idShort, that.idShort)
+                && Objects.equals(endpoints, that.endpoints)
+                && Objects.equals(administration, that.administration)
+                && Objects.equals(descriptions, that.descriptions)
+                && Objects.equals(identification, that.identification);
+    }
+
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(idShort, endpoints, administration, descriptions, identification);
+    }
+
+}
